Experimentation/AttGAN/requirements.txt
absl-py==0.11.0
astor==0.8.1
bleach==1.5.0
cached-property==1.5.2
cycler==0.10.0
gast==0.4.0
grpcio==1.35.0
h5py==2.10.0
html5lib==0.9999999
imageio==2.9.0
importlib-metadata==3.4.0
joblib==1.0.1
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.19.5
pandas==0.25.3
patsy==0.5.1
Pillow==8.1.0
plotly==4.14.3
protobuf==3.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
statsmodels==0.12.1
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.56.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.0
